requirements.txt
cycler==0.11.0
fonttools==4.29.1
kiwisolver==1.3.2
matplotlib==3.5.1
numpy==1.22.2
opencv-python==4.5.5.62
packaging==21.3
pdf2image==1.16.0
Pillow==9.0.1
pyparsing==3.0.7
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.1
pytesseract==0.3.9
python-dateutil==2.8.2
six==1.16.0
tk==0.1.0
